feat: Add InvoiceNotFoundError class
This commit adds the InvoiceNotFoundError class to the domain/errors folder. This class extends the Error class and is used to throw an error when an invoice is not found. The error message includes the ID of the missing invoice.

refactor: Move constants to common folder

This commit moves the constants to the common folder. This helps to organize the code and make it more maintainable.

refactor: Move dtos to application folder

This commit moves the dtos (data transfer objects) to the application folder. This improves the code structure and separates the concerns of the different layers of the application.

feat: Add InvoiceCreatedEvent class

This commit adds the InvoiceCreatedEvent class to the domain/events folder. This class represents an event that is triggered when an invoice is created. It contains the created invoice as a property.

feat: Add MongoDB implementation for invoice repository interface and repository methods according to DDD architecture

This commit adds the MongoDB implementation for the invoice repository interface and repository methods. This implementation follows the Domain-Driven Design (DDD) architecture principles.